changes int() to float()

+++ set_sv.py
@@ -6,6 +6,6 @@ import sys
 instrument = omegacn7500.OmegaCN7500(settings.port,settings.rimsaddressint)
 # port name, slave address
 
-instrument.set_setpoint(int(sys.argv[1]))
+instrument.set_setpoint(float(sys.argv[1]))
 print(sys.argv[1])
 # Registernumber, value, number of decimals for storage
